Added 'spacewait.m" integrator
spacewait is a generic function for waiting a specified amount of time
in space.  I could do this analytically, but it's much easier just doing
this, for now.

In this case, it's being used to wait until the circularization burn.
The wait duration was precalculated using Kepler's equations.  Doing the
calculations with real math taught me a lot about orbital mechanics - I
should implement things analytically when I can, and avoid using
numerical methods out of convenience.