Proven to work!
I am amazed.  The result of MATLAB outputs a rather non-conventional set
of ascent trajectory parameters as the optimal set for this rocket. Just
tried it in KSP itself, and it truly does perform better than the ranges
I expected to be best.

It has disproven a law I held to be true: Low-TWR rockets should have
shallow ascents, and high TWR rockets should have steep ascents.  I had
a low-TWR rocket need a turn shape of around 60%, and a turnstart of
7411 m